+++ 20news/KaggleWord2VecUtility.py
