+++ Reuters/Word2Vec.py
@@ -4,7 +4,7 @@ import nltk.data
 import logging
 import numpy as np
 from gensim.models import Word2Vec
-from KaggleWord2VecUtility_Dheeraj import KaggleWord2VecUtility
+from KaggleWord2VecUtility import KaggleWord2VecUtility
 import time
 from sklearn.preprocessing import normalize
 import sys
