+++ Reuters/metrics.py
@@ -2,7 +2,7 @@ import time;
 import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
 import numpy as np
-from KaggleWord2VecUtility_Dheeraj import KaggleWord2VecUtility
+from KaggleWord2VecUtility import KaggleWord2VecUtility
 from numpy import float32
 import math
 from sklearn.ensemble import RandomForestClassifier
